chore: rename cli module to __main__ so it can be executed without having to build
